requirements.txt
scikit-learn==1.4.1.post1
duckdb
pandas
textattack
textblob
pyspellchecker
torch
transformers
